feat(csv): Add basic csv to JSON conversion
This adds a script `csv_to_json` which is essentially a wrapper for
`csvjson` found in the python3-csvkit package.  This is the first step
for JSON results in Zathras, but for now this is intended to convert the
current CSV files to a results JSON (metadata is stripped) so we can
easily verify those results.

Co-developed-by: Cursor