refactor(detect_os): Remove uname options for detect_os
Removes uname options for detect_os, so uname can no longer be used to
try and determine the OS or it's version.  Usage of uname
detection was non-existent.

os-release was the previous default and now is the only mechanism that
detect_os uses.  Since uname based detection was removed, the dependency
that this script had on `yq` is no longer needed and now it is a
standalone script.

The only thing I found as an issue is amazon linux missing getopt in
its container image.  I doubt this is a problem within a regular ec2
instance.