feat(pkg_tool): Add SLES support
Add SLES support for installing packages.  Due to zypper install dumping
to stdout, this should be blocked by PR #101.  Without that PR, zypper
install dumps to stdout which is not caught and is passed directly to
package_tool, which causes the first run on any system missing `wget` to
fail.

Blocked-by: #101