feat(pkg_tool): Add ability for wrapper_config to specify OS versions
Adds the ability to specify dependencies on an OS version level.  The
previous way of working with dependencies is still supported ("osname":
[<list of packages]) and will work as expected.  This adds support for
checking the full version, major version, and then a default set of
packages for an OS.  An example is below
```json
"rhel": {
  "9.6": ["gcc"],
  "9": ["g++"],
  "default": ["gawk"]
}
```

That will install gcc on 9.6, install g++ on 9.* versions (except 9.6),
and then gawk on all versions (except 9.*).